committing work

+++ lib/ui/dashboard/pharm_dashboard/pharmacy_view_product_screen.dart
@@ -681,8 +681,7 @@ class PharmacyViewProductScreen extends StatelessWidget {
                                                     ),
                                                     TextView(
                                                       text: formatNaira(
-                                                        o['enlistedPricePerUnit']
-                                                            .toInt(),
+                                                        o['enlistedPricePerUnit'],
                                                       ),
                                                       textStyle: TextStyle(
                                                         fontFamily: 'DMSans',
